refactor: move Wine process snapshot helper
