+++ app/src/main/java/app/gamenative/ui/screen/xserver/XServerScreen.kt
@@ -106,6 +106,7 @@ import app.gamenative.utils.ManifestComponentHelper
 import app.gamenative.utils.PreInstallSteps
 import app.gamenative.utils.SteamTokenLogin
 import app.gamenative.utils.SteamUtils
+import app.gamenative.utils.WineProcessSnapshotHelper
 import com.posthog.PostHog
 import com.winlator.alsaserver.ALSAClient
 import com.winlator.container.Container
@@ -281,110 +282,6 @@ private fun buildEssentialProcessAllowlist(): Set<String> {
     val essentialServices = WineUtils.getEssentialServiceNames()
         .map { normalizeProcessName(it) }
     return (essentialServices + CORE_WINE_PROCESSES).toSet()
-}
-
-private fun readWineProcessSnapshotFromProc(): List<ProcessInfo> {
-    val procDir = File("/proc")
-    val pidDirs = try {
-        procDir.listFiles { file -> file.isDirectory && file.name.all { it.isDigit() } } ?: return emptyList()
-    } catch (_: Exception) {
-        return emptyList()
-    }
-
-    val myUid = android.os.Process.myUid()
-    val processes = mutableListOf<ProcessInfo>()
-
-    for (pidDir in pidDirs) {
-        try {
-            val pid = pidDir.name.toIntOrNull() ?: continue
-            if (pid == android.os.Process.myPid()) continue
-
-            val statusLines = try {
-                File(pidDir, "status").readLines()
-            } catch (_: Exception) {
-                emptyList()
-            }
-            val uid = statusLines
-                .firstOrNull { it.startsWith("Uid:") }
-                ?.trim()
-                ?.split(Regex("\\s+"))
-                ?.getOrNull(1)
-                ?.toIntOrNull()
-            if (uid == null || uid != myUid) continue
-
-            val comm = try { File(pidDir, "comm").readText().trim() } catch (_: Exception) { "" }
-            val stat = try { File(pidDir, "stat").readText() } catch (_: Exception) { "" }
-            val cmdlineBytes = try { File(pidDir, "cmdline").readBytes() } catch (_: Exception) { ByteArray(0) }
-            val cmdlineArgs = cmdlineBytes.toNullSeparatedStrings()
-            val searchableText = buildString {
-                append(comm)
-                append(' ')
-                append(stat)
-                append(' ')
-                append(cmdlineArgs.joinToString(" "))
-            }
-
-            if (!searchableText.contains("wine", ignoreCase = true) &&
-                !searchableText.contains(".exe", ignoreCase = true)) {
-                continue
-            }
-
-            val name = cmdlineArgs
-                .firstOrNull { it.endsWith(".exe", ignoreCase = true) }
-                ?: cmdlineArgs.firstOrNull { it.contains(".exe", ignoreCase = true) }
-                ?: cmdlineArgs.firstOrNull { it.contains("wine", ignoreCase = true) }
-                ?: comm.ifBlank { null }
-                ?: continue
-
-            val rssBytes = statusLines
-                .firstOrNull { it.startsWith("VmRSS:") }
-                ?.trim()
-                ?.split(Regex("\\s+"))
-                ?.getOrNull(1)
-                ?.toLongOrNull()
-                ?.times(1024L)
-                ?: 0L
-
-            processes.add(
-                ProcessInfo(
-                    pid,
-                    name.substringAfterLast('/').substringAfterLast('\\'),
-                    rssBytes,
-                    0,
-                    false,
-                ),
-            )
-        } catch (_: Exception) {
-            // Process exited or became unreadable while enumerating /proc.
-        }
-    }
-
-    val allowlist = buildEssentialProcessAllowlist()
-    return processes.sortedWith(
-        compareByDescending<ProcessInfo> { normalizeProcessName(it.name) !in allowlist }
-            .thenByDescending { it.memoryUsage },
-    )
-}
-
-private fun ByteArray.toNullSeparatedStrings(): List<String> {
-    if (isEmpty()) return emptyList()
-
-    val result = mutableListOf<String>()
-    var start = 0
-    for (i in indices) {
-        if (this[i] == 0.toByte()) {
-            if (i > start) {
-                val value = String(this, start, i - start).trim()
-                if (value.isNotBlank()) result.add(value)
-            }
-            start = i + 1
-        }
-    }
-    if (start < size) {
-        val value = String(this, start, size - start).trim()
-        if (value.isNotBlank()) result.add(value)
-    }
-    return result
 }
 
 // TODO logs in composables are 'unstable' which can cause recomposition (performance issues)
@@ -1006,7 +903,7 @@ fun XServerScreen(
         quickMenuWineProcessesLoading = true
         while (showQuickMenu && quickMenuToolsVisible) {
             quickMenuWineProcesses = withContext(Dispatchers.IO) {
-                readWineProcessSnapshotFromProc()
+                WineProcessSnapshotHelper.readFromProc()
             }
             quickMenuWineProcessesLoading = false
             delay(QUICK_MENU_PROCESS_POLL_INTERVAL_MS)
